Refactoring and package changes, changed map style, added cancel button to racecourse input activity

+++ app/src/main/java/com/aayaffe/sailingracecoursemanager/general/GeneralUtils.java
@@ -51,11 +51,11 @@ public class GeneralUtils {
      * @param context Context to get resources and device specific display metrics
      * @return A float value to represent px equivalent to dp depending on device density
      */
-    public static float convertDpToPixel(float dp, Context context){
+    public static int convertDpToPixel(float dp, Context context){
         Resources resources = context.getResources();
         DisplayMetrics metrics = resources.getDisplayMetrics();
         float px = dp * ((float)metrics.densityDpi / DisplayMetrics.DENSITY_DEFAULT);
-        return px;
+        return (int)px;
     }
 
     public static int getDeviceWidth(Activity activity) {
